+++ src/TrashLib/Config/Settings/SettingsValues.cs
@@ -3,6 +3,8 @@ namespace TrashLib.Config.Settings;
 public record TrashRepository
 {
     public string CloneUrl { get; init; } = "https://github.com/TRaSH-/Guides.git";
+    public string Branch { get; init; } = "master";
+    public string? Sha1 { get; init; }
 }
 
 public record SettingsValues