+++ src/VersionControl/IGitRepository.cs
@@ -4,6 +4,6 @@ public interface IGitRepository : IDisposable
 {
     void ForceCheckout(string branch);
     void Fetch(string remote = "origin");
-    void ResetHard(string toBranch);
+    void ResetHard(string toBranchOrSha1);
     void SetRemote(string name, string newUrl);
 }